+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Query Log Viewer:
 
 1. Fix displaying the log viewer when opening a new project into the IDE window.
 1. Don't clear the log viewer when an error is from a previously selected query processor.
+1. Update the server list when the query processors are modified.
 
 ## 2020